Add a new metric, Percentage of time spent in active CPU. It will collect per thread, and report in the thread shutdown.  It will also accumualte as an average, and that average will be reported at test completion.
Add a new command to the main logic. "PauseEvery" is a list of numeric values. We will pause after the number specified, cycling through the list of numbers. The purpose of this is to better represent the web environment, where there may be time when no thread is active. This goes to the time spent in Init/Term.

Split the switch to call the specific worker thread out from the outerthread. This was done to more easily allow the next change.

Run each worker thread inside a try/catch, and if APDFL is active, inside a DURING/HANDLER as well. This will catch any uncaught exceptions thrown by the workers, and assure that we always properly shut down the library.

Signed-off-by: Michael McKeough <mjm@datalogics.com>