Major refactoring.
I split each of the worker threads out into it's own .ccp/.h files. This will allow the designer of a new thread class to simply copy an existing one, and modify it. It also makes reading any particular bit of code a lot simpler. I updated and expanded comments in the process. The base worker class is also split to it's own module.

The enumeration of worker class types is moved to the base worker class header. Each object will now carry aninternal marker of it's type, and will report that marker via GetClassType().

This takes the main module down from 2K lines to 600.

Progress reporting is made optional on the threads that use it. It's display is now properly condition on silent, and logFile.

The default_Commands.txt file is updated to properly report all commands.

I added commands to set paths to the binaries, plugins, resources, and sample data, but have not yet added effectors for these.

Signed-off-by: Michael McKeough <mjm@datalogics.com>